:page_facing_up: Update license notice in changed files
