+++ CPAC/utils/__init__.py
@@ -20,7 +20,6 @@ from .utils import (
     correlation,
     check,
     check_random_state,
-    cl_strip_brackets,
     try_fetch_parameter,
     get_scan_params,
     get_tr,
