+++ CPAC/utils/monitoring/custom_logging.py
@@ -1,6 +1,6 @@
 """Funtions for logging.
 
-Copyright (C) 2022  C-PAC Developers
+Copyright (C) 2022-2023  C-PAC Developers
 
 This file is part of C-PAC.
