+++ CPAC/utils/utils.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2012-2022  C-PAC Developers
+# Copyright (C) 2012-2023  C-PAC Developers
 
 # This file is part of C-PAC.
 
@@ -117,30 +117,7 @@ def check_prov_for_motion_tool(prov):
             return None
 
 
-def cl_strip_brackets(arg_list):
-    """Removes '[' from before first and ']' from after final
-    arguments in a list of commandline arguments
 
-    Parameters
-    ----------
-    arg_list : list
-
-    Returns
-    -------
-    list
-
-    Examples
-    --------
-    >>> cl_strip_brackets('[a b c]'.split(' '))
-    ['a', 'b', 'c']
-    >>> cl_strip_brackets('a b c'.split(' '))
-    ['a', 'b', 'c']
-    >>> cl_strip_brackets('[ a b c ]'.split(' '))
-    ['a', 'b', 'c']
-    """
-    arg_list[0] = arg_list[0].lstrip('[')
-    arg_list[-1] = arg_list[-1].rstrip(']')
-    return [arg for arg in arg_list if arg]
 
 
 def get_flag(in_flag):
